Migrate FirepowerModifier from float to integer percentage.

+++ OpenRA.Game/Traits/TraitsInterfaces.cs
@@ -151,7 +151,7 @@ namespace OpenRA.Traits
 	public interface IRenderModifier { IEnumerable<IRenderable> ModifyRender(Actor self, WorldRenderer wr, IEnumerable<IRenderable> r); }
 	public interface IDamageModifier { int GetDamageModifier(Actor attacker, WarheadInfo warhead); }
 	public interface ISpeedModifier { int GetSpeedModifier(); }
-	public interface IFirepowerModifier { float GetFirepowerModifier(); }
+	public interface IFirepowerModifier { int GetFirepowerModifier(); }
 	public interface ILoadsPalettes { void LoadPalettes(WorldRenderer wr); }
 	public interface IPaletteModifier { void AdjustPalette(IReadOnlyDictionary<string, MutablePalette> b); }
 	public interface IPips { IEnumerable<PipType> GetPips(Actor self); }
